pyproject.toml
[project]
name = "tosca_api"
version = "0.1.0"
description = "TOSCA-Web API - Backend for geospatial web applications"
requires-python = ">=3.12"
readme = "README.md"
authors = [
    { name = "TOSCA Team", email = "team@example.com" }
]
license = "MIT"
dependencies = [
    "django>=5.1,<5.2",
    "djangorestframework>=3.15.0",
    "django-environ>=0.11.0",
    "python-keycloak>=4.0.0",
    "pyjwt[crypto]>=2.8.0",
    "psycopg[binary]>=3.2.0",
    "django-allauth>=65.13.1",
    "dj-rest-auth>=7.0.1",
    "ruff>=0.14.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tosca_api.settings.test"
python_files = "tests.py test_*.py *_tests.py"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.setuptools.packages.find]
include = ["tosca_api*"]
exclude = ["docker*"]
